Add DynamicLearnedIndex baseline comparison and analysis
- Add baseline/ (Sgelet/DynamicLearnedIndex reference implementation)
- Add baseline_wrapper with __int128 BigQuotient for numerical safety
- Add compare.cpp for head-to-head LSLR vs DLI benchmarking
- Add analysis.md with full algorithm descriptions and experimental results
- Add .gitignore for build artifacts
- Update README with correct OptimalPLA algorithm (convex hull, not polygon clipping)
- Update README with DLI comparison section and project structure
- Fix OptimalPLA header: data-dependent bounds, cross-product precision

+++ include/common.hpp
@@ -33,8 +33,8 @@ struct PLASegment {
     size_t end;         // end rank in logical array
     Key first_key;
     Key last_key;
-    double slope;
-    double intercept;
+    long double slope;
+    long double intercept;
 
     // For lazy shift tracking
     int64_t lazy_shift{0};
